requirements.txt
Django==2.2.1
djangorestframework==3.9.4
djangorestframework-csv
rest-pandas
psycopg2==2.8.2
ptvsd==4.2.9
pylint==2.3.1
django-environ
Faker